+++ storm_analysis/test/run_tests.py
@@ -21,25 +21,32 @@ test_args = [
     [dao_exe, "test.dax", "test_3d_2d.bin", "test_3d_2d.xml"],
     [dao_exe, "test.dax", "test_3d_3d.bin", "test_3d_3d.xml"],
     [dao_exe, "test.dax", "test_3d_Z.bin", "test_3d_Z.xml"],
-    ["Testing FRC"],
-    ["../frc/frc_calc2d.py", "test_drift_mlist.bin", "test_drift_frc.txt", "0"],
+
+    #["Testing FRC"],
+    #["../frc/frc_calc2d.py", "test_drift_mlist.bin", "test_drift_frc.txt", "0"],
+
     ["Testing L1H"],
     ["../L1H/setup_A_matrix.py", "theoritical", "test_l1h", "1.0", "0"],
     ["../L1H/cs_analysis.py", "test_l1h.dax", "test_l1h.xml", "test_l1h.hres", "test_l1h_list.bin"],
-    ["Testing RCC"],
-    ["../rcc/rcc-drift-correction.py", "test_drift_mlist.bin", "test_drift.txt", "2000", "1"],
+
+    #["Testing RCC"],
+    #["../rcc/rcc-drift-correction.py", "test_drift_mlist.bin", "test_drift.txt", "2000", "1"],
+
     ["Testing Rolling Ball Background Subtraction"],
     ["../rolling_ball_bgr/rolling_ball.py", "test_bg_sub.dax", "test_bg_sub_rb.dax", "10", "1"],
+
     ["Testing sCMOS"],
     [scmos_exe, "test.dax", "test_sc_2d_fixed.bin", "test_sc_2d_fixed.xml"],
     [scmos_exe, "test.dax", "test_sc_2d.bin", "test_sc_2d.xml"],
     [scmos_exe, "test.dax", "test_sc_3d.bin", "test_sc_3d.xml"],
     [scmos_exe, "test.dax", "test_sc_Z.bin", "test_sc_Z.xml"],
+
     ["Testing Spliner"],
     ["../spliner/measure_psf.py", "test_spliner.dax", "none", "test_spliner_olist.bin", "test_spliner_psf.psf", "1"],
     ["../spliner/psf_to_spline.py", "test_spliner_psf.psf", "test_spliner_psf.spline", "10"],
     [spliner_exe, "test_spliner.dax", "test_spliner_slist.bin", "test_spliner_dh.xml"],
     [spliner_exe, "test_spliner.dax", "test_spliner_flist.bin", "test_spliner_dh_fista.xml"],
+
     ["Testing Track/Average/Correct"],
     ["../sa_utilities/track_average_correct.py", "test_drift_mlist.bin", "test_drift_alist.bin", "test_drift.xml"],
     
@@ -47,9 +54,11 @@ test_args = [
     ["Testing DBSCAN Clustering"],
     ["../dbscan/dbscan_analysis.py", "test_drift_alist.bin", "0"],
     ["../dbscan/cluster_images.py", "test_drift_aclusters_list.bin", "DBSCAN Clustering", "50", "20", "test_drift_dbscan"],
+
     ["Testing Voronoi Clustering"],
     ["../voronoi/voronoi_analysis.py", "test_drift_alist.bin", "1.25", "./"],
     ["../dbscan/cluster_images.py", "test_drift_asrt_list.bin", "Voronoi Clustering", "50", "20", "test_drift_vr"],
+
     ["Testing Wavelet Background Subtraction"],
     ["../wavelet_bgr/wavelet_bgr.py", "test_bg_sub.dax", "test_bg_sub_wbgr.dax", "db4", "2", "2", "10"]
 ]
